Issue26 Delete employee

+++ Museum-Backend/src/main/java/ca/mcgill/ecse321/museum/service/EmployeeService.java
@@ -8,23 +8,33 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import ca.mcgill.ecse321.museum.dao.EmployeeRepository;
+import ca.mcgill.ecse321.museum.dao.ScheduleOfTimePeriodRepository;
 import ca.mcgill.ecse321.museum.model.Employee;
+import ca.mcgill.ecse321.museum.model.ScheduleOfTimePeriod;
 
 @Service
 public class EmployeeService {
 
   @Autowired
-  EmployeeRepository employeeRepository;  
+  EmployeeRepository employeeRepository;
+
+  @Autowired
+  ScheduleOfTimePeriodRepository scheduleOfTimePeriodRepository;
 
   /**
    * Method to get an employee by their id
    * 
-   * @param id - long
+   * @param id - Long
    * @return employee
    * @author Siger
    */
   @Transactional
-  public Employee getEmployee(long id) {
+  public Employee getEmployee(Long id) {
+    // Error handling
+    if (id == null) {
+      throw new IllegalArgumentException("Employee id cannot be null");
+    }
+
     return employeeRepository.findEmployeeByMuseumUserId(id);
   }
 
@@ -44,12 +54,24 @@ public class EmployeeService {
    * Method to delete an employee from the database by their id
    * Allows the manager to delete an employee
    * 
-   * @param id - long
+   * @param id - Long
    * @return if the employee was deleted (success)
    * @author Siger
    */
   @Transactional
-  public boolean deleteEmployee(long id) {
+  public boolean deleteEmployee(Long id) {
+    // Check if the employee exists and error handling
+    Employee employee = employeeRepository.findEmployeeByMuseumUserId(id);
+    if (employee == null) {
+      throw new IllegalArgumentException("Employee does not exist");
+    }
+
+    // Delete scheduleOfTimePeriods
+    ScheduleOfTimePeriod scheduleOfTimePeriod = scheduleOfTimePeriodRepository.findScheduleOfTimePeriodBySchedule(employee.getSchedule());
+    if (scheduleOfTimePeriod != null) {
+      scheduleOfTimePeriodRepository.deleteScheduleOfTimePeriodBySchedule(employee.getSchedule());
+    }
+
     // Delete employee
     employeeRepository.deleteEmployeeByMuseumUserId(id);
